Implement a classpath scanner
The approach used here is to allow modules to request bindings for
interfaces to implementations found by scanning the class path, as
long as only one such implementation is found. This allows modules to
bind by requesting bindings for the main interface(s) they want to
provide, and re-use the annotation-based information contained in
their implementations and their dependencies.

Signed-off-by: Stephen Kitt <skitt@redhat.com>